+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Improve `Ethers.deploy/2` error handling
 - Implement `Ethers.CcipRead` to support EIP-3668
+- NameService improvements and offchain lookup support using CCIP-Read
 
 ## v0.5.5 (2024-12-03)